In 117

+++ src/features/edit-profile/lib/hooks/useProfileAvatar.ts
@@ -1,6 +1,7 @@
 import { useState } from 'react'
 
 import { profileApi } from '@/common/api/profile.api'
+import { useConfirmation } from '@/common/ui/action-confirmation/useConfirmation'
 
 export const useProfileAvatar = () => {
   const [isOpen, setIsOpen] = useState(false)
@@ -9,6 +10,8 @@ export const useProfileAvatar = () => {
   const [setAvatarProfile, { isLoading: isLoadingSet }] = profileApi.useSetAvatarProfileMutation()
   const [deleteAvatarProfile, { isLoading: isLoadingDelete }] =
     profileApi.useDeleteAvatarProfileMutation()
+  const { confirmOpen, handleConfirm, handleReject, requestConfirmation, setConfirmOpen } =
+    useConfirmation()
 
   const isLoading = isProfileLoading || isLoadingSet || isLoadingDelete
   const currentAvatar = profile?.avatars?.[0] || null
@@ -25,6 +28,13 @@ export const useProfileAvatar = () => {
     if (!currentAvatar) {
       return
     }
+
+    const confirmed = await requestConfirmation()
+
+    if (!confirmed) {
+      return
+    }
+
     try {
       await deleteAvatarProfile().unwrap()
     } catch (error) {
@@ -33,10 +43,14 @@ export const useProfileAvatar = () => {
   }
 
   return {
+    confirmOpen,
     currentAvatar,
+    handleConfirm,
+    handleReject,
     isLoading,
     isOpen,
     removeAvatar,
+    setConfirmOpen,
     setIsOpen,
     updateAvatar,
   }
